Rename is-wp-errors.js to be just errors.js and moved isPermissionScopeError function to it.
