+++ assets/js/googlesitekit/datastore/user/constants.js
@@ -26,6 +26,3 @@ export const PERMISSION_VIEW_DASHBOARD = 'googlesitekit_view_dashboard';
 export const PERMISSION_VIEW_MODULE_DETAILS = 'googlesitekit_view_module_details';
 export const PERMISSION_MANAGE_OPTIONS = 'googlesitekit_manage_options';
 export const PERMISSION_PUBLISH_POSTS = 'googlesitekit_publish_posts';
-
-// Error code returned when scopes are missing.
-export const ERROR_MISSING_REQUIRED_SCOPE = 'missing_required_scopes';
