--- assets/js/googlesitekit/datastore/user/utils/is-permission-scope-error.js
@@ -1,35 +1,0 @@
-/**
- * isPermissionScopeError utility.
- *
- * Site Kit by Google, Copyright 2020 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * Internal dependencies
- */
-import { ERROR_MISSING_REQUIRED_SCOPE } from '../constants';
-
-/**
- * Checks if the given error is a permission scope error.
- *
- * @since 1.9.0
- * @private
- *
- * @param {*} error Input to test as a possible permission scope error.
- * @return {boolean} True if permission scope error, otherwise false.
- */
-export function isPermissionScopeError( error ) {
-	return error?.code === ERROR_MISSING_REQUIRED_SCOPE;
-}
