+++ assets/js/modules/analytics/components/common/AccountCreate/index.js
@@ -29,6 +29,7 @@ import Button from '../../../../../components/button';
 import Link from '../../../../../components/link';
 import ProgressBar from '../../../../../components/progress-bar';
 import { trackEvent } from '../../../../../util';
+import { ERROR_CODE_MISSING_REQUIRED_SCOPE } from '../../../../../util/errors';
 import TimezoneSelect from './TimezoneSelect';
 import AccountField from './AccountField';
 import PropertyField from './PropertyField';
@@ -37,7 +38,7 @@ import CountrySelect from './CountrySelect';
 import StoreErrorNotice from '../../../../../components/StoreErrorNotice';
 import { STORE_NAME, FORM_ACCOUNT_CREATE, PROVISIONING_SCOPE } from '../../../datastore/constants';
 import { STORE_NAME as CORE_SITE } from '../../../../../googlesitekit/datastore/site/constants';
-import { STORE_NAME as CORE_USER, ERROR_MISSING_REQUIRED_SCOPE } from '../../../../../googlesitekit/datastore/user/constants';
+import { STORE_NAME as CORE_USER } from '../../../../../googlesitekit/datastore/user/constants';
 import { STORE_NAME as CORE_FORMS } from '../../../../../googlesitekit/datastore/forms/constants';
 import { countryCodesByTimezone } from '../../../util/countries-timezones';
 import Data from 'googlesitekit-data';
@@ -94,7 +95,7 @@ export default function AccountCreate() {
 				// When state is restored, auto-submit the request again.
 				setValues( FORM_ACCOUNT_CREATE, { autoSubmit: true } );
 				setPermissionScopeError( {
-					code: ERROR_MISSING_REQUIRED_SCOPE,
+					code: ERROR_CODE_MISSING_REQUIRED_SCOPE,
 					message: __( 'Additional permissions are required to create a new Analytics account.', 'google-site-kit' ),
 					data: {
 						status: 403,
