+++ assets/js/modules/analytics/components/setup/SetupForm.js
@@ -44,7 +44,7 @@ import {
 } from '../common';
 import StoreErrorNotice from '../../../../components/StoreErrorNotice';
 import { trackEvent } from '../../../../util';
-import { isPermissionScopeError } from '../../../../googlesitekit/datastore/user/utils/is-permission-scope-error';
+import { isPermissionScopeError } from '../../../../util/errors';
 const { useSelect, useDispatch } = Data;
 
 export default function SetupForm( { finishSetup } ) {
