+++ assets/js/modules/tagmanager/components/setup/SetupForm.js
@@ -30,7 +30,7 @@ import Button from '../../../../components/button';
 import { STORE_NAME, FORM_SETUP, EDIT_SCOPE } from '../../datastore/constants';
 import { STORE_NAME as CORE_FORMS } from '../../../../googlesitekit/datastore/forms/constants';
 import { STORE_NAME as CORE_USER } from '../../../../googlesitekit/datastore/user/constants';
-import { isPermissionScopeError } from '../../../../googlesitekit/datastore/user/utils/is-permission-scope-error';
+import { isPermissionScopeError } from '../../../../util/errors';
 import {
 	AccountSelect,
 	AMPContainerSelect,
